extend max class
