+++ max31865/CMakeLists.txt
@@ -1,7 +1,9 @@
 
     add_library (max31865
-        ${CMAKE_CURRENT_SOURCE_DIR}/src/max31865.cpp    
-        ${CMAKE_CURRENT_SOURCE_DIR}/inc/max31865.h     
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/max31865.cpp  
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/max31865_reg.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/inc/max31865.h  
+        ${CMAKE_CURRENT_SOURCE_DIR}/inc/max31865_reg.h     
     )
     target_include_directories (max31865 PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/inc)
     target_link_libraries(max31865 PUBLIC spi)
